Add tapeffect and gray notes support.

+++ Assets/Codes/InGame/GameCamera.cs
@@ -7,7 +7,7 @@ public class GameCamera : MonoBehaviour
     void Awake()
     {
 
-        float aspectRatio = (float)Screen.width / (float)Screen.height;
+        float aspectRatio = (float)Screen.width / Screen.height;
         float size = aspectRatio >= 16f/9 ? 5 : (-3.9375f * aspectRatio + 12f);
 
         GetComponent<Camera>().orthographicSize = size;
